+++ src/main/java/bots/CalendarBot/Event.java
@@ -7,12 +7,12 @@ package bots.CalendarBot;
  */
 public class Event {
 	protected String wat;
-	protected int date, repeat, remind, nextRepeat, nextRemind;
+	protected int date, repeat, remind, nextRepeat, nextRemind, firstYear;
 	
 	/**
 	 * 
 	 * @param wat description of event
-	 * @param date date of event (minutes since 01.01.2000 00:00)
+	 * @param date date of event (minutes since 01.01.{@code firstYear})
 	 * @param repeat number of days to repeat event (-1 = off)
 	 * @param remind number of minutes to remind before event starts (-1 = off)
 	 * @param currentDate current date as past minutes since 01.01.{@code firstYear} 
@@ -23,43 +23,17 @@ public class Event {
 		this.date = date;
 		this.repeat = repeat;
 		this.remind = remind;
-		setNextRepeat(currentDate, firstYear);
-		setNextRemind(firstYear);
+		this.firstYear = firstYear;
+		setNextRepeat(currentDate);
+		setNextRemind();
 	}
 	
-	
-	public void setWat (String wat) {
-		this.wat = wat;
-	}
-	
-	public void setDate (int date) {
-		this.date = date;
-	}
-	
-	public void setRepeat (int repeat) {
-		this.repeat = repeat;
-	}
-	
-	public void setRemind (int remind) {
-		this.remind = remind;
-	}
-	
-	public void setNextRepeat (int currentDate, int firstYear) {
+	private void setNextRepeat (int currentDate) {
 		nextRepeat = date;
 		int[] nextRepeatSplitted = CalDateFormat.minToDateSplitted(date, firstYear);
 		int correctDay = nextRepeatSplitted[0];
 		while (nextRepeat < currentDate && repeat > 0) {
-			if (repeat % 360 == 0) {
-				nextRepeatSplitted[0] = correctDay;
-				for (int i = 0; i < repeat / 360; i++) {					
-					nextRepeatSplitted[2] += 1;
-				}
-				while (!CalDateFormat.checkDate(nextRepeatSplitted[0], nextRepeatSplitted[1], nextRepeatSplitted[2], nextRepeatSplitted[3], nextRepeatSplitted[4]) && nextRepeatSplitted[0] > 0) {
-					nextRepeatSplitted[0]--;
-				}
-				nextRepeat = CalDateFormat.dateToMin(CalDateFormat.dateSplittedToDate(nextRepeatSplitted[0], nextRepeatSplitted[1], nextRepeatSplitted[2], nextRepeatSplitted[3], nextRepeatSplitted[4]), firstYear);				
-			}
-			else if (repeat % 30 == 0) {
+			if (repeat % 30 == 0) {
 				nextRepeatSplitted[0] = correctDay;
 				for (int i = 0; i < repeat / 30; i++) {
 					nextRepeatSplitted[1] += 1;
@@ -68,6 +42,7 @@ public class Event {
 						nextRepeatSplitted[2] += 1; 
 					}
 				}
+				//check if date exist (e.g: 31.4 do not exists) -if not existing: decrease day
 				while (!CalDateFormat.checkDate(nextRepeatSplitted[0], nextRepeatSplitted[1], nextRepeatSplitted[2], nextRepeatSplitted[3], nextRepeatSplitted[4]) && nextRepeatSplitted[0] > 0) {
 					nextRepeatSplitted[0]--;
 				}
@@ -79,31 +54,19 @@ public class Event {
 		}
 	}
 	
-	public void setNextRemind (int firstYear) {		
+	private void setNextRemind () {		
 		nextRemind = nextRepeat;
 		int[] nextRemindSplitted = CalDateFormat.minToDateSplitted(nextRepeat, firstYear);
-		int correctDay = nextRemindSplitted[0];
-		//525600 = minutes of a year
-		if (remind % 525600 == 0 && remind > 0) {
-			nextRemindSplitted[0] = correctDay;
-			for (int i = 0; i < remind / 525600; i++) {
-				nextRemindSplitted[2] += -1;
-			}
-			while (!CalDateFormat.checkDate(nextRemindSplitted[0], nextRemindSplitted[1], nextRemindSplitted[2], nextRemindSplitted[3], nextRemindSplitted[4]) && nextRemindSplitted[0] > 0) {
-				nextRemindSplitted[0]--;
-			}
-			nextRemind = CalDateFormat.dateToMin(CalDateFormat.dateSplittedToDate(nextRemindSplitted[0], nextRemindSplitted[1], nextRemindSplitted[2], nextRemindSplitted[3], nextRemindSplitted[4]), firstYear);
-		}
 		//43200 = minutes of a month
-		else if (remind % 43200 == 0 && remind > 0) {
-			nextRemindSplitted[0] = correctDay;
+		if (remind % 43200 == 0 && remind > 0) {
 			for (int i = 0; i < remind / 43200; i++) {				
 				nextRemindSplitted[1] += -1;
 				if (nextRemindSplitted[1] == 0) {
 					nextRemindSplitted[1] = 12;
 					nextRemindSplitted[2] += -1; 
 				}
-			}
+			}				
+			//check if date exist (e.g: 31.4 do not exists) - if not existing: decrease day
 			while (!CalDateFormat.checkDate(nextRemindSplitted[0], nextRemindSplitted[1], nextRemindSplitted[2], nextRemindSplitted[3], nextRemindSplitted[4]) && nextRemindSplitted[0] > 0) {
 				nextRemindSplitted[0]--;
 			}
@@ -157,12 +120,8 @@ public class Event {
 		return date + " " + wat + " " + repeat + " " + remind;
 	}
 	
-	public String toStringList (int firstYear) {
-		return CalDateFormat.minToDate(date, firstYear) + ": " + wat + ", repeat: " + repeatToString() + ", remind: " + remindToString();
-	}
-	
-	public String toStringNext (int firstYear) {
-		return CalDateFormat.minToDate(nextRepeat, firstYear) + ": " + wat + ", repeat: " + repeatToString() + ", remind: " + remindToString();
+	public String toStringList (boolean next) {
+		return CalDateFormat.minToDate((next ? nextRepeat : date), firstYear) + ": " + wat + ", repeat: " + repeatToString() + ", remind: " + remindToString();
 	}
 	
 	/**
@@ -170,7 +129,7 @@ public class Event {
 	 * @param firstYear
 	 * @return reminiscence of event
 	 */
-	public String toStringRemind (int firstYear) {
+	public String toStringRemind () {
 		return CalDateFormat.minToWeekday(nextRepeat, firstYear) + " " + CalDateFormat.minToDate(nextRepeat, firstYear) + ": " + wat;
 	}
 	
@@ -179,7 +138,7 @@ public class Event {
 	 * @param firstYear
 	 * @return
 	 */
-	public String toStringRepeat (int firstYear) {
+	public String toStringRepeat () {
 		return "Today " + CalDateFormat.minToDate(nextRepeat, firstYear).substring(11, 16) + ": " + wat;
 	}
 	
@@ -197,8 +156,8 @@ public class Event {
 	}
 	
 	protected String remindToString () {
-		if (remind % 525600 == 0) {
-			return "" + (remind / 525600) + (remind == 525600 ? " year" : " years");
+		if (remind % 518400 == 0) {
+			return "" + (remind / 518400) + (remind == 518400 ? " year" : " years");
 		}
 		if (remind % 43200 == 0) {
 			return "" + (remind / 43200) + (remind == 43200 ? " month" : " months");