+++ src/main/java/bots/CalendarBot/Changelog.md
@@ -2,19 +2,38 @@
 
 v. 1.0.6:
 ---------
-Bugfixes: Alex Bug
+- Bugfixes: Alex Bug
 
 Alex Bug: command eventls shows "repeat" and "remind" of events as number of days/ minutes
 
 
 v. 1.0.7:
 ---------
-Bugfixes: Next Bug
+- Bugfixes: Next Bug
 
 Next Bug: command "eventls next" shows the oldest repeating event instead of the next repeating event
 
 v. 1.0.8:
 ---------
-Bugifxes: S-Bug
+- Bugifxes: S-Bug
 
 S-Bug: names of person in Birthday calendar which end with s or x are listed wrongly
+
+v. 1.1.0:
+--------
+- new:
+	- command "version" (now visible)
+	- age-related birthday wishes (ages 18, 20, others)
+	- new birthday wishes for ages 18 and 20
+
+- Refactoring
+
+- Multiple Bugfixes
+
+Bugs:
+	- command "eventls [calendar] next" shows wrong date
+	- Error "unknown command" shows wrong wrong command
+
+
+
+